// dotnet/openapi/model/applicantModel.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApplicantModel { 
    fullName?: string | null;
    email?: string | null;
    phone?: string | null;
    resumeUrl?: string | null;
    skills?: string | null;
    experience?: number;
    jobId?: string;
    id?: string;
    clientCompanyName?: string | null;
    jobTitle?: string | null;
    appliedDate?: string;
    lastUpdatedDate?: string;
    status?: number;
    statusName?: string | null;
}