// dotnet/openapi/model/getAllApplicantsQuery.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetAllApplicantsQuery { 
    pageIndex?: number;
    pageSize?: number;
    searchString?: string | null;
    clientCompanyId?: string | null;
    jobId?: string | null;
}